kinet-trainer/requirements.txt
eel>=0.16.0
Pillow>=9.0
numpy>=1.21
torch>=1.10
scikit-image>=0.19
scipy>=1.7
